requirements.txt
beautifulsoup4
openai
pandas
pydantic
requests
streamlit>=1.23.0
google-generativeai